Released V3.4.1
[2020-10-29]
1. Refined dynamic classloader issue for springboot.
2. Enhanced full class name for compiled rules to avoid the issue of duplicated xml file name. FullClassNameMode is enabled by default.
   Using 'com.unibeta.vrules.tools.CommonSyntaxs.setEnableFullClassNameMode(boolean)' to enable or disable FullClassName mode.
3. Enhanced rule sequence support double data type.
4. Released V3.4.1.

+++ src/main/java/com/unibeta/vrules/parsers/RulesParser.java
@@ -950,7 +950,7 @@ public class RulesParser {
 		}
 
 		String seqStr = ruleElement.getAttribute(SEQUENCE).trim();
-		long sequence = Long.valueOf(null == seqStr || "".equals(seqStr) ? 0 : Long.valueOf(seqStr)).longValue();
+		double sequence = Double.valueOf(null == seqStr || "".equals(seqStr) ? 0 : Double.valueOf(seqStr)).longValue();
 
 		String brakpointStr = ruleElement.getAttribute(BRAKPOINT).toUpperCase().trim();
 		String dependsStr = ruleElement.getAttribute(DEPENDS).trim();
